overrides the styles refactor the edit text box strokes
