+++ app/src/main/java/com/tailorfit/android/tailorfitapp/dashboard/DashBoardFragment.kt
@@ -1,0 +1,22 @@
+package com.tailorfit.android.tailorfitapp.dashboard
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.tailorfit.android.R
+import com.tailorfit.android.base.BaseFragment
+
+
+class DashBoardFragment : BaseFragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_dash_board, container, false)
+    }
+
+}
