+++ app/src/main/res/layout/fragment_sign_in.xml
@@ -27,11 +27,11 @@
             android:id="@+id/sign_in_here_text"
             android:layout_width="fill_parent"
             android:layout_height="wrap_content"
+            android:layout_marginTop="16dp"
             android:fontFamily="@font/lato_bold"
             android:text="@string/sign_in_here"
             android:textColor="@color/textColorPrimaryGrey"
             android:textSize="16sp"
-            android:layout_marginTop="16dp"
             app:layout_constraintBottom_toTopOf="@id/_phone_edit_text_container"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
@@ -50,6 +50,7 @@
 
             <com.google.android.material.textfield.TextInputLayout
                 android:id="@+id/phoneTextInputLayout"
+                style="@style/TextInputLayoutStyle"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:hint="@string/phone_number">
@@ -76,6 +77,7 @@
 
             <com.google.android.material.textfield.TextInputLayout
                 android:id="@+id/passwordTextInputLayout"
+                style="@style/TextInputLayoutStyle"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:hint="@string/password">
@@ -88,7 +90,6 @@
 
             </com.google.android.material.textfield.TextInputLayout>
         </FrameLayout>
-
 
 
         <com.google.android.material.button.MaterialButton
