+++ app/src/main/res/values-v23/styles.xml
@@ -3,6 +3,5 @@
 
     <style name="AppTheme" parent="AppThemeParent">
         <item name="android:windowLightStatusBar">false</item>
-
     </style>
 </resources>
