+++ app/src/main/res/values/styles.xml
@@ -1,6 +1,6 @@
 <resources>
 
-    <style name="AppTheme" parent="AppThemeParent"/>
+    <style name="AppTheme" parent="AppThemeParent" />
 
     <style name="AppTheme.HorizontalButtonAppearance" parent="Widget.MaterialComponents.Button.Icon">
         <item name="android:fontFamily">@font/lato_bold</item>
@@ -11,6 +11,15 @@
         <item name="android:layout_marginEnd">20dp</item>
         <item name="android:textColor">@android:color/white</item>
         <item name="backgroundTint">@color/secondaryColorChocolate</item>
+    </style>
+
+
+    <style name="TextInputLayoutStyle" parent="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.Dense">
+        <item name="boxStrokeWidth">3dp</item>
+        <item name="boxStrokeColor">@color/textColorLightGrey</item>
+        <item name="cornerRadius">8dp</item>
+        <item name="android:textColorHint">@color/textColorPrimaryGrey</item>
+
     </style>
